Intermediate commit for analysis framework.
- Change filenames: "pulse" -> "img_pulse"
- frame counters should be size_t, not long
- add username to analysis summary
- img_pulse.py can do both all pixels and individual pixels
- disable FITS/HDF5 stuff in Makefiles
- fix img_pulse_test
- mostly fix display of img_pulse results;
    need to accomodate multiple modules, and all-pixel output

TODO: graphs of img_pulse results.
I plan to switch from D3 (fancy Javascript thing) to good old gnuplot.
We need to handle extremely large (TB) result files.

+++ analysis/img_pulse.cpp
@@ -32,7 +32,7 @@
 #include <math.h>
 
 #include "pff.h"
-#include "pulse_find.h"
+#include "img_pulse.h"
 
 #define MAX_VAL             65536        // ignore values larger than this
 
@@ -72,7 +72,7 @@ inline double sample_to_sec(long i) {
 
 // called when a pulse is complete
 //
-void PULSE_FIND::pulse_complete(int level, double value, long isample) {
+void PULSE_FIND::pulse_complete(int level, double value, size_t isample) {
     int idur = 1<<level;
     isample = isample + 1 - idur;
     if (isample < 0) return;
